update progres kategori, resep, lokasi seller

+++ app/Http/Controllers/Api/SatuanController.php
@@ -3,8 +3,12 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use App\Models\Aset;
+use App\Models\Resep;
 use Illuminate\Http\Request;
 use App\Models\Satuan;
+use App\Models\UserLog;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Validator;
 
 class SatuanController extends Controller
@@ -12,7 +16,7 @@ class SatuanController extends Controller
     public function index()
     {
         // with(['kategori', 'stok']) itu Eager Loading biar gak lemot
-        $satuans = Satuan::where(['aktif' => true])->orderBy('nama', 'asc')->get();
+        $satuans = Satuan::where(['aktif' => true])->orderBy('id', 'asc')->get();
 
         return response()->json([
             'success' => true,
@@ -40,14 +44,23 @@ class SatuanController extends Controller
             ], 422); // 422 Unprocessable Entity
         }
 
-        // Simpan ke database
-        $satuan = Satuan::create($request->all());
+        try {
+            DB::transaction(function () use ($request) {
+                $satuan = Satuan::create($request->all());
+                UserLog::simpan("Menambah satuan baru dengan nama {$satuan->nama}");
+            });
 
-        return response()->json([
-            'success' => true,
-            'message' => 'Data Satuan berhasil ditambahkan',
-            'data'    => $satuan
-        ], 201); // 201 Created
+            return response()->json([
+                'success' => true,
+                'message' => 'Data Satuan berhasil ditambahkan'
+            ], 201); // 201 Created
+        } catch (\Exception $th) {
+            return response()->json([
+                'success'  => false,
+                'message' => 'Gagal menyimpan satuan. Transaksi dibatalkan.',
+                'error'   => $th->getMessage()
+            ], 500);
+        }
     }
 
     public function show($id)
@@ -95,19 +108,32 @@ class SatuanController extends Controller
             ], 422);
         }
 
-        // Jalankan update
-        $satuan->update($request->all());
+        try {
+            DB::transaction(function () use ($request, $satuan) {
+                $namaLama = $satuan->nama;
+                // Jalankan update
+                $satuan->update($request->all());
+                UserLog::simpan("Mengubah satuan dari {$namaLama} menjadi {$satuan->nama}", ["semula" => $namaLama, "menjadi" => $satuan->nama]);
+            });
 
-        return response()->json([
-            'success' => true,
-            'message' => 'Data Satuan berhasil diupdate',
-            'data'    => $satuan
-        ], 200);
+            return response()->json([
+                'success' => true,
+                'message' => 'Data Satuan berhasil diupdate'
+            ], 200);
+        } catch (\Exception $th) {
+            return response()->json([
+                'success'  => false,
+                'message' => 'Gagal menyimpan satuan. Transaksi dibatalkan.',
+                'error'   => $th->getMessage()
+            ], 500);
+        }
     }
 
     public function destroy($id)
     {
         $satuan = Satuan::find($id);
+        $asetBySatuan = Aset::where(['satuan_id' => $id])->orWhere(['satuan_ecer_id' => $id])->get();
+        $resepBySatuan = Resep::where(['satuan_id' => $id])->get();
 
         if (!$satuan) {
             return response()->json([
@@ -116,11 +142,30 @@ class SatuanController extends Controller
             ], 404);
         }
 
-        $satuan->delete();
+        if ($asetBySatuan || $resepBySatuan) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Masih ada Aset/Resep yang menggunakan Satuan ini'
+            ], 404);
+        }
 
-        return response()->json([
-            'success' => true,
-            'message' => 'Data Satuan berhasil dihapus'
-        ], 200);
+        try {
+            DB::transaction(function () use ($satuan) {
+                $namaLama = $satuan->nama;
+                $satuan->delete();
+                UserLog::simpan("Menghapus satuan {$namaLama}");
+            });
+
+            return response()->json([
+                'success' => true,
+                'message' => 'Data Satuan berhasil dihapus'
+            ], 200);
+        } catch (\Exception $th) {
+            return response()->json([
+                'success'  => false,
+                'message' => 'Gagal menghapus satuan. Transaksi dibatalkan.',
+                'error'   => $th->getMessage()
+            ], 500);
+        }
     }
 }
